+++ src/types/biometry/consent.ts
@@ -1,0 +1,4 @@
+export interface ConsentResponse {
+  is_consent_given: boolean;
+  user_fullname: string;
+}
