+++ src/types/biometry/face-match.ts
@@ -1,0 +1,13 @@
+export interface FaceMatchResponse {
+  code: number;
+  result: number;
+  description: string;
+  anchor: {
+    code: number;
+    description: string;
+  };
+  target: {
+    code: number;
+    description: string;
+  };
+}
