+++ src/types/biometry/session.ts
@@ -1,0 +1,4 @@
+export interface SessionResponse {
+  data: string;
+  message: string;
+}
