+++ src/types/internal.ts
@@ -1,0 +1,4 @@
+export interface ApiResponse<T> {
+  body: T;
+  headers: Record<string, string>;
+}
